PW-32: handle page token for search queries
